+++ noodles-tabix/src/io.rs
@@ -1,3 +1,6 @@
 //! Tabix I/O.
 
 pub mod indexed_reader;
+mod reader;
+
+pub use self::reader::Reader;