+++ noodles-tabix/src/lib.rs
@@ -30,10 +30,12 @@ pub mod r#async;
 
 pub mod index;
 pub mod io;
-mod reader;
 mod writer;
 
-pub use self::{reader::Reader, writer::Writer};
+#[deprecated(since = "0.45.0", note = "Use `noodles_tabix::io::Reader` instead.")]
+pub use self::io::Reader;
+
+pub use self::writer::Writer;
 
 #[cfg(feature = "async")]
 pub use self::r#async::{Reader as AsyncReader, Writer as AsyncWriter};